backend/gemini_backend/requirements.txt
google-generativeai==0.3.2
opencv-python==4.9.0.80
